Tag source-image-build-utils image with git sha
Previously, we tagged the image only with the 'latest' tag.

In quay.io, an image that doesn't have any tags cannot be pulled. By
pushing to the 'latest' tag, we would immediately invalidate the
previous image, breaking everyone who relied on the image.

Tag the image with the git commit sha to make sure every image will
always have at least one tag associated with it.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>