Create sbom-utility-scripts image
This image contains multiple python scripts that are used in the buildah
tasks. This is done so that multiple steps in buildah tasks can be
merged. The image contains scripts from the following stages:

merge-syft-sboms
merge-cachi2-sbom
create-purl-sbom
create-base-images-sbom

NOTE: the "merge_syft_sbom" from the "merge-cachi2-sbom" stage was taken
from
https://github.com/containerbuildsystem/cachi2/blob/main/utils/merge_syft_sbom.py
and renamed to "merge_cachi2_sboms.py". It will be removed from the
cachi2 repo and image.

STONEBLD-2608

Signed-off-by: mkosiarc <mkosiarc@redhat.com>