Put all prefetched dependencies of one kind to one layer
This changes how we are creating layers for source builds. Instead of
each dependency being in an individual layer, all dependencies for
specific package manager will be added to a single layer.

e.g. all npm dependencies will be in one layer and all
pip dependencies in another one.

The purpose of this is to limit the number of layers.

KFLUXBUGS-1508

Signed-off-by: mkosiarc <mkosiarc@redhat.com>