Improve generation of base images SBOMs
The process how the base image SBOM is generated was slightly changed.

The most important change is due to requiring different inputs. The
script now requires the full json of parsed Dockerfile, so we can
extract the base images better. This allows us to extract base images
data even for stage aliases.

Another advantage is that we are counting
the stages/layers more carefully, because we have more information about
which stage was from scratch or from oci-archive.

KFLUXBUGS-1718

Signed-off-by: mkosiarc <mkosiarc@redhat.com>