merge_sboms: implement merge_spdx_sboms function
Merges the packages of both SBOMs in exactly the same way as we do for
CycloneDX.

Merges the creationInfo by concatenating and deduplicating the lists of
creators of both documents.

Merges the relationships by concatenating and processing as follows:

- if the SPDXID on either side of the relationship:
  - is the ID of SBOM B -> replace it with the ID of SBOM A
  - is not the ID of SBOM A or the ID of any package in the
    merged package list -> drop the whole relationship
  - otherwise -> keep it

Drops the .files list because we don't care about merging it.

Note that the merging does not have special handling for the
"root packages" of either document (the packages in relationship
DOCUMENT_ID DESCRIBES ROOT_ID). That means the merged document can have
two or more roots. This should be fine, especially because the
add_image_reference.py script in this repo is already able to handle
that.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>