merge_sboms: fall back to deduplicating by ID
When merging packages that don't have a purl, we used to deduplicate by
{name}@{version}. This can be problematic for SPDX SBOMs, which may use
a name-less, version-less package as the "root":

    {
        "SPDXID": "SPDXRef-MadeUpRoot-A",
        "name": "",
        "versionInfo": ""
    }

    SPDXRef-Document DESCRIBES SPDXRef-MadeUpRoot-A
    SPDXRef-MadeUpRoot-A CONTAINS <most of the packages in this SBOM>

Given an SPDXRef-MadeUpRoot-A and SPDXRef-MadeUpRoot-B, both with no
name and version, our previous approach would end up dropping
SPDXRef-MadeUpRoot-B. This would also result in dropping all the
CONTAINS and other relationships involving SPDXRef-MadeUpRoot-B.

For packages that have neither a purl nor a name, fall back to the
"id" - the thing that uniquely identifies a package within an SBOM. For
SPDX packages, this is the SPDXID attribute. For CycloneDX components,
it is the bom-ref attribute (which is optional, but hopefully we will
never see a component with none of (purl, name, bom-ref)).

This reduces the chance that we will drop packages we shouldn't drop.
If two "made up root" packages have the same SPDXID, that is fine - one
of them will be dropped, but all relationships will be preserved
(because the SPDXID they reference still exists). And the fact that we
dropped the package shouldn't be a problem in itself - it had no purl
and no name, so it probably carried 0 relevant information.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>